chore: move endpoint integration methods to a file
