+++ e2e/_suites/fleet/endpoint.go
@@ -1,0 +1,286 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
+package main
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/cenkalti/backoff/v4"
+	"github.com/cucumber/godog"
+	"github.com/elastic/e2e-testing/internal/common"
+	"github.com/elastic/e2e-testing/internal/deploy"
+	"github.com/elastic/e2e-testing/internal/utils"
+	log "github.com/sirupsen/logrus"
+)
+
+func (fts *FleetTestSuite) theIntegrationIsOperatedInThePolicy(packageName string, action string) error {
+	return theIntegrationIsOperatedInThePolicy(fts.currentContext, fts.kibanaClient, fts.Policy, packageName, action)
+}
+
+func (fts *FleetTestSuite) theHostNameIsNotShownInTheAdminViewInTheSecurityApp() error {
+	log.Trace("Checking if the hostname is not shown in the Administration view in the Security App")
+
+	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
+	retryCount := 1
+
+	exp := utils.GetExponentialBackOff(maxTimeout)
+
+	agentListedInSecurityFn := func() error {
+		agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
+		manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
+		host, err := fts.kibanaClient.IsAgentListedInSecurityApp(fts.currentContext, manifest.Hostname)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"elapsedTime": exp.GetElapsedTime(),
+				"err":         err,
+				"host":        host,
+				"hostname":    manifest.Hostname,
+				"retry":       retryCount,
+			}).Warn("We could not check the agent in the Administration view in the Security App yet")
+
+			retryCount++
+
+			return err
+		}
+
+		log.WithFields(log.Fields{
+			"elapsedTime": exp.GetElapsedTime(),
+			"hostname":    manifest.Hostname,
+			"retries":     retryCount,
+		}).Info("The Agent is not listed in the Administration view in the Security App")
+		return nil
+	}
+
+	err := backoff.Retry(agentListedInSecurityFn, exp)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (fts *FleetTestSuite) theHostNameIsShownInTheAdminViewInTheSecurityApp(status string) error {
+	log.Trace("Checking if the hostname is shown in the Admin view in the Security App")
+
+	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
+	retryCount := 1
+
+	exp := utils.GetExponentialBackOff(maxTimeout)
+
+	agentListedInSecurityFn := func() error {
+		agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
+		manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
+		matches, err := fts.kibanaClient.IsAgentListedInSecurityAppWithStatus(fts.currentContext, manifest.Hostname, status)
+		if err != nil || !matches {
+			log.WithFields(log.Fields{
+				"elapsedTime":   exp.GetElapsedTime(),
+				"desiredStatus": status,
+				"err":           err,
+				"hostname":      manifest.Hostname,
+				"matches":       matches,
+				"retry":         retryCount,
+			}).Warn("The agent is not listed in the Administration view in the Security App in the desired status yet")
+
+			retryCount++
+
+			return err
+		}
+
+		log.WithFields(log.Fields{
+			"elapsedTime":   exp.GetElapsedTime(),
+			"desiredStatus": status,
+			"hostname":      manifest.Hostname,
+			"matches":       matches,
+			"retries":       retryCount,
+		}).Info("The Agent is listed in the Administration view in the Security App in the desired status")
+		return nil
+	}
+
+	err := backoff.Retry(agentListedInSecurityFn, exp)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (fts *FleetTestSuite) thePolicyIsUpdatedToHaveMode(name string, mode string) error {
+	if name != "malware" {
+		log.WithFields(log.Fields{
+			"name": name,
+		}).Warn("We only support 'malware' policy to be updated")
+		return godog.ErrPending
+	}
+
+	if mode != "detect" && mode != "prevent" {
+		log.WithFields(log.Fields{
+			"name": name,
+			"mode": mode,
+		}).Warn("We only support 'detect' and 'prevent' modes")
+		return godog.ErrPending
+	}
+
+	packageDS, err := fts.kibanaClient.GetIntegrationFromAgentPolicy(fts.currentContext, "endpoint", fts.Policy)
+
+	if err != nil {
+		return err
+	}
+	fts.Integration = packageDS.Package
+
+	for _, item := range packageDS.Inputs {
+		if item.Type == "endpoint" {
+			item.Config.(map[string]interface{})["policy"].(map[string]interface{})["value"].(map[string]interface{})["windows"].(map[string]interface{})["malware"].(map[string]interface{})["mode"] = mode
+			item.Config.(map[string]interface{})["policy"].(map[string]interface{})["value"].(map[string]interface{})["mac"].(map[string]interface{})["malware"].(map[string]interface{})["mode"] = mode
+		}
+	}
+	log.WithFields(log.Fields{
+		"inputs": packageDS.Inputs,
+	}).Trace("Upgrading integration package config")
+
+	updatedAt, err := fts.kibanaClient.UpdateIntegrationPackagePolicy(fts.currentContext, packageDS)
+	if err != nil {
+		return err
+	}
+
+	// we use a string because we are not able to process what comes in the event, so we will do
+	// an alphabetical order, as they share same layout but different millis and timezone format
+	fts.PolicyUpdatedAt = updatedAt
+	return nil
+}
+
+func (fts *FleetTestSuite) thePolicyWillReflectTheChangeInTheSecurityApp() error {
+	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
+	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
+	agentID, err := fts.kibanaClient.GetAgentIDByHostname(fts.currentContext, manifest.Hostname)
+	if err != nil {
+		return err
+	}
+
+	pkgPolicy, err := fts.kibanaClient.GetIntegrationFromAgentPolicy(fts.currentContext, "endpoint", fts.Policy)
+	if err != nil {
+		return err
+	}
+
+	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute * 2
+	retryCount := 1
+
+	exp := utils.GetExponentialBackOff(maxTimeout)
+
+	getEventsFn := func() error {
+		err := fts.kibanaClient.GetAgentEvents(fts.currentContext, "endpoint-security", agentID, pkgPolicy.ID, fts.PolicyUpdatedAt)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"elapsedTime": exp.GetElapsedTime(),
+				"err":         err,
+				"retries":     retryCount,
+			}).Warn("There are no events for the agent in Fleet")
+			retryCount++
+
+			return err
+		}
+
+		log.WithFields(log.Fields{
+			"elapsedTime": exp.GetElapsedTime(),
+			"retries":     retryCount,
+		}).Info("There are events for the agent in Fleet")
+		return nil
+	}
+
+	err = backoff.Retry(getEventsFn, exp)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (fts *FleetTestSuite) thePolicyResponseWillBeShownInTheSecurityApp() error {
+	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName)
+	manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
+	agentID, err := fts.kibanaClient.GetAgentIDByHostname(fts.currentContext, manifest.Hostname)
+	if err != nil {
+		return err
+	}
+
+	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
+	retryCount := 1
+
+	exp := utils.GetExponentialBackOff(maxTimeout)
+
+	getEventsFn := func() error {
+		listed, err := fts.kibanaClient.IsPolicyResponseListedInSecurityApp(fts.currentContext, agentID)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"elapsedTime": exp.GetElapsedTime(),
+				"err":         err,
+				"retries":     retryCount,
+			}).Warn("Could not get metadata from the Administration view in the Security App yet")
+			retryCount++
+
+			return err
+		}
+
+		if !listed {
+			log.WithFields(log.Fields{
+				"agentID":     agentID,
+				"elapsedTime": exp.GetElapsedTime(),
+				"retries":     retryCount,
+			}).Warn("The policy response is not listed as 'success' in the Administration view in the Security App yet")
+			retryCount++
+
+			return fmt.Errorf("the policy response is not listed as 'success' in the Administration view in the Security App yet")
+		}
+
+		log.WithFields(log.Fields{
+			"elapsedTime": exp.GetElapsedTime(),
+			"retries":     retryCount,
+		}).Info("The policy response is listed as 'success' in the Administration view in the Security App")
+		return nil
+	}
+
+	err = backoff.Retry(getEventsFn, exp)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (fts *FleetTestSuite) thePolicyShowsTheDatasourceAdded(packageName string) error {
+	log.WithFields(log.Fields{
+		"policyID": fts.Policy.ID,
+		"package":  packageName,
+	}).Trace("Checking if the policy shows the package added")
+
+	maxTimeout := time.Minute
+	retryCount := 1
+
+	exp := utils.GetExponentialBackOff(maxTimeout)
+
+	configurationIsPresentFn := func() error {
+		packagePolicy, err := fts.kibanaClient.GetIntegrationFromAgentPolicy(fts.currentContext, packageName, fts.Policy)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"packagePolicy": packagePolicy,
+				"policy":        fts.Policy,
+				"retry":         retryCount,
+				"error":         err,
+			}).Warn("The integration was not found in the policy")
+			retryCount++
+			return err
+		}
+
+		retryCount++
+		return err
+	}
+
+	err := backoff.Retry(configurationIsPresentFn, exp)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
